dashboard_control: carve api_sessions.rs (pure move)
