+++ src/bin/caller/dashboard_control/api_sessions.rs
@@ -1,0 +1,1641 @@
+//! Session-scoped control requests: request/stream task spawns, session
+//! lists and streams, detail/report/history, current-session uploads,
+//! rollback/redo/prune/changes, context snapshots, and session search.
+
+use super::*;
+
+pub(crate) fn spawn_control_request(
+    id: String,
+    method: String,
+    params: Option<serde_json::Value>,
+    runtime: ControlRuntime,
+    task_tx: mpsc::Sender<ControlTaskResponse>,
+    pending_requests: &mut HashMap<String, CancellationToken>,
+) {
+    if let Some(previous) = pending_requests.remove(&id) {
+        previous.cancel();
+    }
+    let cancel = CancellationToken::new();
+    pending_requests.insert(id.clone(), cancel.clone());
+    tokio::spawn(async move {
+        let response = match method.as_str() {
+            "api_session_report" => {
+                api_session_report_task_response(id.clone(), params.as_ref(), &runtime).await
+            }
+            "api_session_current_upload_raw" => {
+                api_session_current_upload_raw_task_response(id.clone(), params.as_ref(), &runtime)
+                    .await
+            }
+            "api_recording_asset" => {
+                api_recording_asset_task_response(id.clone(), params.as_ref(), &runtime).await
+            }
+            "api_session_recording_asset" => {
+                api_session_recording_asset_task_response(id.clone(), params.as_ref()).await
+            }
+            "api_session_frame_asset" => {
+                api_session_frame_asset_task_response(id.clone(), params.as_ref()).await
+            }
+            "api_fs_read" => api_fs_read_task_response(id.clone(), params.as_ref()).await,
+            "api_transfer_download_read" => {
+                api_transfer_download_read_task_response(id.clone(), params.as_ref(), &runtime)
+                    .await
+            }
+            _ => {
+                let frame =
+                    control_request_response(id.clone(), method, params, runtime, cancel).await;
+                ControlTaskResponse {
+                    id,
+                    frame,
+                    byte_stream: None,
+                    done: true,
+                }
+            }
+        };
+        let _ = task_tx.send(response).await;
+    });
+}
+
+pub(crate) fn spawn_control_stream(
+    id: String,
+    method: String,
+    params: Option<serde_json::Value>,
+    task_tx: mpsc::Sender<ControlTaskResponse>,
+    pending_requests: &mut HashMap<String, CancellationToken>,
+) {
+    if let Some(previous) = pending_requests.remove(&id) {
+        previous.cancel();
+    }
+    let cancel = CancellationToken::new();
+    pending_requests.insert(id.clone(), cancel.clone());
+    tokio::spawn(async move {
+        match method.as_str() {
+            "api_sessions_stream" => {
+                stream_sessions_response(id, params.as_ref(), task_tx, cancel).await;
+            }
+            _ => {
+                let frame = serde_json::json!({
+                    "t": "stream_end",
+                    "id": id,
+                    "ok": false,
+                    "error": format!("unknown stream method: {method}"),
+                });
+                let _ = task_tx
+                    .send(ControlTaskResponse {
+                        id,
+                        frame,
+                        byte_stream: None,
+                        done: true,
+                    })
+                    .await;
+            }
+        }
+    });
+}
+
+/// Test the client-egress path end to end: force a tiny provider call
+/// through the relay for `kind` — even when a local key or lease exists —
+/// and return the model's reply. The fueling panel's "test relay" button
+/// and the E2E validator's deterministic hook.
+pub(crate) async fn api_credential_egress_probe_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let kind = params
+        .and_then(|p| optional_string_param(p, &["kind"]))
+        .unwrap_or_default();
+    let model = params.and_then(|p| optional_string_param(p, &["model"]));
+    let provider: Box<dyn crate::provider::ChatProvider> = match kind.as_str() {
+        crate::credential_egress::KIND_ANTHROPIC => {
+            Box::new(crate::provider::AnthropicProvider::new_client_egress(
+                model.unwrap_or_else(|| "claude-haiku-4-5-20251001".to_string()),
+                200_000,
+                256,
+            ))
+        }
+        crate::credential_egress::KIND_GEMINI => {
+            Box::new(crate::provider::GeminiProvider::new_client_egress(
+                model.unwrap_or_else(|| "gemini-2.5-flash".to_string()),
+                200_000,
+                256,
+            ))
+        }
+        other => {
+            return dashboard_control_error_response(
+                id,
+                format!(
+                    "egress probe supports {} and {}, not {other:?}",
+                    crate::credential_egress::KIND_ANTHROPIC,
+                    crate::credential_egress::KIND_GEMINI
+                ),
+            )
+        }
+    };
+    let probe_message = crate::conversation::Message {
+        role: "user".to_string(),
+        content: "Reply with the single word: pong".to_string(),
+        ..Default::default()
+    };
+    match provider.chat(&[probe_message]).await {
+        Ok(response) => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": true,
+            "result": { "text": response.content, "model": provider.model() },
+        }),
+        Err(error) => dashboard_control_error_response(id, format!("egress probe failed: {error}")),
+    }
+}
+
+pub(crate) async fn control_request_response(
+    id: String,
+    method: String,
+    params: Option<serde_json::Value>,
+    runtime: ControlRuntime,
+    cancel: CancellationToken,
+) -> serde_json::Value {
+    if cancel.is_cancelled() {
+        return cancelled_control_response(id, true);
+    }
+    match method.as_str() {
+        "api_credential_egress_probe" => {
+            api_credential_egress_probe_response(id, params.as_ref()).await
+        }
+        "api_access_connect_unclaim" => {
+            match crate::web_gateway::access_connect_unclaim_response_value(
+                runtime.project_root.clone(),
+            )
+            .await
+            {
+                Ok(result) => serde_json::json!({
+                    "t": "response",
+                    "id": id,
+                    "ok": true,
+                    "result": result,
+                }),
+                Err(error) => serde_json::json!({
+                    "t": "response",
+                    "id": id,
+                    "ok": false,
+                    "error": error,
+                }),
+            }
+        }
+        "api_sessions" => api_sessions_response(id, params.as_ref()).await,
+        "api_session_detail" => api_session_detail_response(id, params.as_ref()).await,
+        "api_session_delete" => api_session_delete_response(id, params.as_ref()).await,
+        "api_session_agent_output" => api_session_agent_output_response(id, params.as_ref()).await,
+        "api_session_current_agent_output" => {
+            api_session_current_agent_output_response(id, params.as_ref(), &runtime).await
+        }
+        "api_session_current_history" => api_session_current_history_response(id, &runtime).await,
+        "api_session_current_rollback" => {
+            api_session_current_rollback_response(id, params.as_ref(), &runtime).await
+        }
+        "api_session_current_redo" => api_session_current_redo_response(id, &runtime).await,
+        "api_session_current_prune" => api_session_current_prune_response(id, &runtime).await,
+        "api_session_current_changes" => {
+            api_session_current_changes_response(id, params.as_ref(), &runtime).await
+        }
+        "api_session_context_snapshot" => {
+            api_session_context_snapshot_response(id, params.as_ref()).await
+        }
+        "api_session_current_uploads" => api_session_current_uploads_response(id, &runtime).await,
+        "api_session_current_upload_delete" => {
+            api_session_current_upload_delete_response(id, params.as_ref(), &runtime).await
+        }
+        "api_transfer_jobs" => api_transfer_jobs_response(id, &runtime).await,
+        "api_transfer_job_create" => {
+            api_transfer_job_create_response(id, params.as_ref(), &runtime).await
+        }
+        "api_transfer_job_delete" => {
+            api_transfer_job_delete_response(id, params.as_ref(), &runtime).await
+        }
+        "api_transfer_upload_commit" => {
+            api_transfer_upload_commit_response(id, params.as_ref(), &runtime).await
+        }
+        "api_media_clip_start" => {
+            api_media_clip_start_response(id, params.as_ref(), &runtime).await
+        }
+        "api_media_clip_end" => api_media_clip_end_response(id, params.as_ref(), &runtime).await,
+        "api_media_clip_cancel" => {
+            api_media_clip_cancel_response(id, params.as_ref(), &runtime).await
+        }
+        "api_fs_stat" => api_fs_stat_response(id, params.as_ref()).await,
+        "api_fs_list" => api_fs_list_response(id, params.as_ref()).await,
+        "api_fs_mkdir" => api_fs_mkdir_response(id, params.as_ref()).await,
+        "api_fs_rename" => api_fs_rename_response(id, params.as_ref()).await,
+        "api_fs_delete" => api_fs_delete_response(id, params.as_ref()).await,
+        "api_sessions_search" => api_sessions_search_response(id, params.as_ref(), cancel).await,
+        "api_settings" => api_settings_response(id, &runtime).await,
+        "api_settings_save" => api_settings_save_response(id, params.as_ref(), &runtime).await,
+        "api_control_msg" => api_control_msg_response(id, params.as_ref(), &runtime).await,
+        "api_session_control_msg" => {
+            api_session_control_msg_response(id, params.as_ref(), &runtime).await
+        }
+        "api_dashboard_action_msg" => {
+            api_dashboard_action_msg_response(id, params.as_ref(), &runtime).await
+        }
+        "api_diagnostics_visual_freshness" => {
+            api_diagnostics_visual_freshness_response(id, params.as_ref()).await
+        }
+        "api_key_status" => json_body_response(
+            id,
+            crate::web_gateway::api_key_status_response_body(),
+            "api key status",
+        ),
+        "api_external_agents" => json_body_response(
+            id,
+            crate::web_gateway::external_agents_response_body(runtime.project_root.as_deref()),
+            "external agents",
+        ),
+        "api_api_keys_save" => api_api_keys_save_response(id, params.as_ref()).await,
+        "api_voice_session" => api_voice_session_response(id, &runtime).await,
+        "api_project_root" => json_body_response(
+            id,
+            crate::web_gateway::project_root_response_body(runtime.project_root.as_deref()),
+            "project root",
+        ),
+        "api_displays" => api_displays_response(id, &runtime).await,
+        "api_recordings" => api_recordings_response(id, &runtime).await,
+        "api_session_recordings" => api_session_recordings_response(id, params.as_ref()).await,
+        "api_browser_workspace_snapshot" => api_browser_workspace_snapshot_response(id).await,
+        "api_state_snapshot" => api_state_snapshot_response(id, &runtime).await,
+        "api_display_bootstrap" => api_display_bootstrap_response(id, &runtime).await,
+        "api_display_webrtc_signal" => {
+            api_display_webrtc_signal_response(id, params.as_ref(), &runtime).await
+        }
+        "api_display_input_authority_snapshot" => {
+            api_display_input_authority_snapshot_response(id, &runtime).await
+        }
+        "api_display_input_authority_request" => {
+            api_display_input_authority_request_response(id, params.as_ref(), &runtime).await
+        }
+        "api_display_input_authority_release" => {
+            api_display_input_authority_release_response(id, params.as_ref(), &runtime).await
+        }
+        "api_session_log_replay" => api_session_log_replay_response(id, &runtime).await,
+        "api_external_session_activity_replay" => {
+            api_external_session_activity_replay_response(id, &runtime).await
+        }
+        "api_dashboard_bootstrap" => api_dashboard_bootstrap_response(id, &runtime).await,
+        "api_worktrees" => api_worktrees_response(id, &runtime).await,
+        "api_worktrees_inspect" => {
+            api_worktrees_inspect_response(id, params.as_ref(), &runtime).await
+        }
+        "api_worktrees_scan" => api_worktrees_scan_response(id, &runtime).await,
+        "api_worktrees_remove" => {
+            api_worktrees_remove_response(id, params.as_ref(), &runtime).await
+        }
+        "api_managed_context_records" => {
+            api_managed_context_response(id, "records", params.as_ref(), &runtime).await
+        }
+        "api_managed_context_anchors" => {
+            api_managed_context_response(id, "anchors", params.as_ref(), &runtime).await
+        }
+        "api_managed_context_fission" => {
+            api_managed_context_response(id, "fission", params.as_ref(), &runtime).await
+        }
+        "api_mcp_tool_call" => api_mcp_tool_call_response(id, params.as_ref(), &runtime).await,
+        "api_peer_add" => api_peer_add_response(id, params.as_ref(), &runtime).await,
+        "api_peer_remove" => api_peer_remove_response(id, params.as_ref(), &runtime).await,
+        "api_peer_eligible" => api_peer_eligible_response(id, params.as_ref(), &runtime).await,
+        "api_peer_message" => api_peer_message_response(id, params.as_ref(), &runtime).await,
+        "api_peer_task" => api_peer_task_response(id, params.as_ref(), &runtime).await,
+        "api_peer_approval" => api_peer_approval_response(id, params.as_ref(), &runtime).await,
+        "api_peer_webrtc_signal" => {
+            api_peer_webrtc_signal_response(id, params.as_ref(), &runtime).await
+        }
+        "api_peer_file_transfer_signal" => {
+            api_peer_file_transfer_signal_response(id, params.as_ref(), &runtime).await
+        }
+        "api_peer_dashboard_control_signal" => {
+            api_peer_dashboard_control_signal_response(id, params.as_ref(), &runtime).await
+        }
+        "api_peer_pairing_invite" => api_peer_pairing_invite_response(id, params.as_ref()).await,
+        "api_peer_pairing_join" => {
+            api_peer_pairing_join_response(id, params.as_ref(), &runtime).await
+        }
+        "api_peer_pairing_request_access" => {
+            api_peer_pairing_request_access_response(id, params.as_ref()).await
+        }
+        "api_peer_pairing_request_access_poll" => {
+            api_peer_pairing_request_access_poll_response(id, params.as_ref(), &runtime).await
+        }
+        "api_peer_pairing_requests" => api_peer_pairing_requests_response(id).await,
+        "api_peer_pairing_request_decision" => {
+            api_peer_pairing_request_decision_response(id, params.as_ref()).await
+        }
+        "api_peer_pairing_identities" => api_peer_pairing_identities_response(id).await,
+        "api_peer_pairing_identity_revoke" => {
+            api_peer_pairing_identity_revoke_response(id, params.as_ref()).await
+        }
+        "api_coordinator_route" => {
+            api_coordinator_route_response(id, params.as_ref(), &runtime).await
+        }
+        _ => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": format!("unknown method: {method}"),
+        }),
+    }
+}
+
+pub(crate) async fn stream_sessions_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    task_tx: mpsc::Sender<ControlTaskResponse>,
+    cancel: CancellationToken,
+) {
+    let request_line = sessions_stream_request_line(params);
+    let (line_tx, line_rx) = mpsc::channel::<String>(64);
+    let stream_task = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::stream_sessions_from_request(&request_line, line_tx);
+    });
+    stream_json_lines_response(
+        id,
+        "api_sessions_stream".to_string(),
+        line_rx,
+        stream_task,
+        task_tx,
+        cancel,
+    )
+    .await;
+}
+
+pub(crate) async fn stream_json_lines_response(
+    id: String,
+    method: String,
+    mut line_rx: mpsc::Receiver<String>,
+    stream_task: tokio::task::JoinHandle<()>,
+    task_tx: mpsc::Sender<ControlTaskResponse>,
+    cancel: CancellationToken,
+) {
+    if cancel.is_cancelled() {
+        return;
+    }
+
+    if task_tx
+        .send(ControlTaskResponse {
+            id: id.clone(),
+            frame: serde_json::json!({
+                "t": "stream_start",
+                "id": id,
+                "method": method,
+            }),
+            byte_stream: None,
+            done: false,
+        })
+        .await
+        .is_err()
+    {
+        return;
+    }
+
+    let mut seq: u64 = 0;
+    while let Some(line) = line_rx.recv().await {
+        if cancel.is_cancelled() {
+            return;
+        }
+        let trimmed = line.trim();
+        if trimmed.is_empty() {
+            continue;
+        }
+        let event = match serde_json::from_str::<serde_json::Value>(trimmed) {
+            Ok(event) => event,
+            Err(e) => {
+                let frame = serde_json::json!({
+                    "t": "stream_end",
+                    "id": id,
+                    "ok": false,
+                    "error": format!("session stream returned invalid JSON: {e}"),
+                });
+                let _ = task_tx
+                    .send(ControlTaskResponse {
+                        id,
+                        frame,
+                        byte_stream: None,
+                        done: true,
+                    })
+                    .await;
+                return;
+            }
+        };
+        let chunk_id = format!("{id}:{seq}");
+        let frame = serde_json::json!({
+            "t": "stream_event",
+            "id": id,
+            "seq": seq,
+            "chunk_id": chunk_id,
+            "event": event,
+        });
+        seq = seq.saturating_add(1);
+        if task_tx
+            .send(ControlTaskResponse {
+                id: id.clone(),
+                frame,
+                byte_stream: None,
+                done: false,
+            })
+            .await
+            .is_err()
+        {
+            return;
+        }
+    }
+
+    let frame = match stream_task.await {
+        Ok(()) => serde_json::json!({
+            "t": "stream_end",
+            "id": id,
+            "ok": true,
+            "result": {
+                "events": seq,
+            },
+        }),
+        Err(e) => serde_json::json!({
+            "t": "stream_end",
+            "id": id,
+            "ok": false,
+            "error": format!("session stream task failed: {e}"),
+        }),
+    };
+    if !cancel.is_cancelled() {
+        let _ = task_tx
+            .send(ControlTaskResponse {
+                id,
+                frame,
+                byte_stream: None,
+                done: true,
+            })
+            .await;
+    }
+}
+
+pub(crate) fn sessions_stream_request_line(params: Option<&serde_json::Value>) -> String {
+    let Some(params) = params else {
+        return "GET /api/sessions/stream HTTP/1.1".to_string();
+    };
+    let Some(limit_value) = params.get("limit") else {
+        return "GET /api/sessions/stream HTTP/1.1".to_string();
+    };
+    let limit = match limit_value {
+        serde_json::Value::String(value) => {
+            let value = value.trim();
+            if value.eq_ignore_ascii_case("all") || value.eq_ignore_ascii_case("full") {
+                "all".to_string()
+            } else {
+                value
+                    .parse::<usize>()
+                    .ok()
+                    .filter(|limit| *limit > 0)
+                    .unwrap_or(CONTROL_DEFAULT_SESSION_LIMIT)
+                    .to_string()
+            }
+        }
+        serde_json::Value::Number(value) => value
+            .as_u64()
+            .and_then(|limit| usize::try_from(limit).ok())
+            .filter(|limit| *limit > 0)
+            .unwrap_or(CONTROL_DEFAULT_SESSION_LIMIT)
+            .to_string(),
+        _ => CONTROL_DEFAULT_SESSION_LIMIT.to_string(),
+    };
+    format!("GET /api/sessions/stream?limit={limit} HTTP/1.1")
+}
+
+pub(crate) fn cancelled_control_response(id: String, existed: bool) -> serde_json::Value {
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": false,
+        "cancelled": true,
+        "error": if existed {
+            "request cancelled"
+        } else {
+            "request not found or already completed"
+        },
+    })
+}
+
+pub(crate) async fn api_session_detail_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let session_id = string_param(&params, &["session_id", "sessionId", "id"]);
+    if session_id.is_empty() {
+        return serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": "missing session_id",
+        });
+    }
+    let source = string_param(&params, &["source"]).trim().to_string();
+    let source = if source.is_empty() {
+        "intendant".to_string()
+    } else {
+        source
+    };
+    let limit = control_session_detail_limit(&params);
+    let before = control_session_detail_before(&params);
+    let body = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::session_detail_response_body_with_page(
+            &session_id,
+            &source,
+            limit,
+            before,
+        )
+    })
+    .await;
+    let body = match body {
+        Ok(body) => body,
+        Err(e) => {
+            return serde_json::json!({
+                "t": "response",
+                "id": id,
+                "ok": false,
+                "error": format!("session detail task failed: {e}"),
+            });
+        }
+    };
+    json_body_response(id, body, "session detail")
+}
+
+pub(crate) async fn api_session_report_task_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> ControlTaskResponse {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let session_id = optional_string_param(&params, &["session_id", "sessionId", "id"])
+        .unwrap_or_else(|| "current".to_string());
+    let (session_log, query_ctx) = {
+        let session = runtime.shared_session.read().await;
+        (session.session_log.clone(), session.query_ctx.clone())
+    };
+    let result = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::session_report_zip_for_request(
+            &session_id,
+            session_log.as_ref(),
+            query_ctx.as_ref(),
+        )
+    })
+    .await;
+    let report = match result {
+        Ok(Ok(report)) => report,
+        Ok(Err(err)) => {
+            let (status, error) = match err {
+                crate::web_gateway::SessionReportZipError::InvalidSessionId => {
+                    (400, "invalid session id".to_string())
+                }
+                crate::web_gateway::SessionReportZipError::NotFound => {
+                    (404, "Session not found".to_string())
+                }
+                crate::web_gateway::SessionReportZipError::Build(error) => {
+                    (500, format!("Failed to build report: {error}"))
+                }
+            };
+            let frame = http_body_response(
+                id.clone(),
+                status,
+                serde_json::json!({
+                    "ok": false,
+                    "error": error,
+                })
+                .to_string(),
+                "session report",
+            );
+            return ControlTaskResponse {
+                id,
+                frame,
+                byte_stream: None,
+                done: true,
+            };
+        }
+        Err(e) => {
+            return ControlTaskResponse {
+                id: id.clone(),
+                frame: serde_json::json!({
+                    "t": "response",
+                    "id": id,
+                    "ok": false,
+                    "error": format!("session report task failed: {e}"),
+                }),
+                byte_stream: None,
+                done: true,
+            };
+        }
+    };
+    let size = report.bytes.len();
+    let filename = report.filename;
+    let content_type = "application/zip".to_string();
+    ControlTaskResponse {
+        id: id.clone(),
+        frame: serde_json::Value::Null,
+        byte_stream: Some(ControlByteStream {
+            id: id.clone(),
+            stream_id: format!("{id}:session-report"),
+            content_type: content_type.clone(),
+            filename: Some(filename.clone()),
+            bytes: report.bytes,
+            result: serde_json::json!({
+                "ok": true,
+                "filename": filename,
+                "content_type": content_type,
+                "size": size,
+            }),
+        }),
+        done: true,
+    }
+}
+
+pub(crate) async fn api_session_current_upload_raw_task_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> ControlTaskResponse {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let Some(upload_id) = optional_string_param(&params, &["id", "upload_id", "uploadId"]) else {
+        return ControlTaskResponse {
+            id: id.clone(),
+            frame: http_body_response(
+                id,
+                400,
+                serde_json::json!({ "ok": false, "error": "missing upload id" }).to_string(),
+                "upload raw",
+            ),
+            byte_stream: None,
+            done: true,
+        };
+    };
+    let offset = match optional_u64_param(&params, &["offset", "start"]) {
+        Ok(offset) => offset.unwrap_or(0),
+        Err(error) => {
+            return ControlTaskResponse {
+                id: id.clone(),
+                frame: http_body_response(
+                    id,
+                    400,
+                    serde_json::json!({ "ok": false, "error": error }).to_string(),
+                    "upload raw",
+                ),
+                byte_stream: None,
+                done: true,
+            };
+        }
+    };
+    let length = match optional_u64_param(&params, &["length", "limit"]) {
+        Ok(length) => length,
+        Err(error) => {
+            return ControlTaskResponse {
+                id: id.clone(),
+                frame: http_body_response(
+                    id,
+                    400,
+                    serde_json::json!({ "ok": false, "error": error }).to_string(),
+                    "upload raw",
+                ),
+                byte_stream: None,
+                done: true,
+            };
+        }
+    };
+    let Some(root) = runtime.project_root.clone() else {
+        return ControlTaskResponse {
+            id: id.clone(),
+            frame: http_body_response(
+                id,
+                404,
+                serde_json::json!({ "ok": false, "error": "no project root" }).to_string(),
+                "upload raw",
+            ),
+            byte_stream: None,
+            done: true,
+        };
+    };
+    let session_log = {
+        let session = runtime.shared_session.read().await;
+        session.session_log.clone()
+    };
+    let session_dir_result = match session_log {
+        Some(ref slog) => slog
+            .lock()
+            .map(|log| log.dir().to_path_buf())
+            .map_err(|_| "session log lock poisoned".to_string()),
+        None => Ok(crate::web_gateway::pending_upload_session_dir(&root)),
+    };
+    let session_dir = match session_dir_result {
+        Ok(session_dir) => session_dir,
+        Err(error) => {
+            return ControlTaskResponse {
+                id: id.clone(),
+                frame: http_body_response(
+                    id,
+                    500,
+                    serde_json::json!({ "ok": false, "error": error }).to_string(),
+                    "upload raw",
+                ),
+                byte_stream: None,
+                done: true,
+            };
+        }
+    };
+    let upload_id_for_stream = upload_id.clone();
+    let read_result = tokio::task::spawn_blocking(move || {
+        let Some(descriptor) = crate::upload_store::find_upload(&upload_id, &session_dir, &root)
+        else {
+            return Err((
+                404,
+                serde_json::json!({ "ok": false, "error": "upload not found" }),
+            ));
+        };
+        let metadata = std::fs::metadata(&descriptor.path).map_err(|e| {
+            (
+                500,
+                serde_json::json!({ "ok": false, "error": format!("stat upload: {e}") }),
+            )
+        })?;
+        let total_size = metadata.len();
+        if offset > total_size {
+            return Err((
+                416,
+                serde_json::json!({
+                    "ok": false,
+                    "error": "range start beyond upload size",
+                    "total_size": total_size,
+                }),
+            ));
+        }
+        let available = total_size.saturating_sub(offset);
+        let requested = length.unwrap_or(available).min(available);
+        if requested > crate::web_gateway::UPLOAD_MAX_BYTES as u64 {
+            return Err((
+                413,
+                serde_json::json!({
+                    "ok": false,
+                    "error": format!(
+                        "range too large: {} bytes (cap is {})",
+                        requested,
+                        crate::web_gateway::UPLOAD_MAX_BYTES
+                    ),
+                }),
+            ));
+        }
+        let transfer_len = usize::try_from(requested).map_err(|_| {
+            (
+                413,
+                serde_json::json!({ "ok": false, "error": "range too large for this platform" }),
+            )
+        })?;
+        let mut file = std::fs::File::open(&descriptor.path).map_err(|e| {
+            (
+                500,
+                serde_json::json!({ "ok": false, "error": format!("open upload: {e}") }),
+            )
+        })?;
+        file.seek(std::io::SeekFrom::Start(offset)).map_err(|e| {
+            (
+                500,
+                serde_json::json!({ "ok": false, "error": format!("seek upload: {e}") }),
+            )
+        })?;
+        let mut bytes = vec![0u8; transfer_len];
+        file.read_exact(&mut bytes).map_err(|e| {
+            (
+                500,
+                serde_json::json!({ "ok": false, "error": format!("read upload: {e}") }),
+            )
+        })?;
+        let end = offset.saturating_add(requested);
+        let descriptor_id = descriptor.id.clone();
+        let descriptor_name = descriptor.name.clone();
+        let descriptor_mime = descriptor.mime.clone();
+        Ok((
+            descriptor_name.clone(),
+            descriptor_mime.clone(),
+            bytes,
+            serde_json::json!({
+                "ok": true,
+                "id": descriptor_id,
+                "name": descriptor_name,
+                "filename": descriptor_name,
+                "mime": descriptor_mime,
+                "content_type": descriptor_mime,
+                "size": requested,
+                "total_size": total_size,
+                "offset": offset,
+                "range_start": offset,
+                "range_end": end,
+                "resumable": true,
+            }),
+        ))
+    })
+    .await;
+    let (filename, content_type, bytes, result) = match read_result {
+        Ok(Ok(value)) => value,
+        Ok(Err((status, body))) => {
+            return ControlTaskResponse {
+                id: id.clone(),
+                frame: http_body_response(id, status, body.to_string(), "upload raw"),
+                byte_stream: None,
+                done: true,
+            };
+        }
+        Err(e) => {
+            return ControlTaskResponse {
+                id: id.clone(),
+                frame: serde_json::json!({
+                    "t": "response",
+                    "id": id,
+                    "ok": false,
+                    "error": format!("upload raw task failed: {e}"),
+                }),
+                byte_stream: None,
+                done: true,
+            };
+        }
+    };
+    ControlTaskResponse {
+        id: id.clone(),
+        frame: serde_json::Value::Null,
+        byte_stream: Some(ControlByteStream {
+            id: id.clone(),
+            stream_id: format!("{id}:upload:{upload_id_for_stream}"),
+            content_type,
+            filename: Some(filename),
+            bytes,
+            result,
+        }),
+        done: true,
+    }
+}
+
+pub(crate) async fn api_session_current_uploads_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let (project_root, session_dir) = match active_upload_handles(runtime).await {
+        Ok(handles) => handles,
+        Err(error) => {
+            return http_body_response(
+                id,
+                500,
+                serde_json::json!({ "error": error }).to_string(),
+                "current uploads",
+            );
+        }
+    };
+    let Some(root) = project_root else {
+        return http_body_response(
+            id,
+            404,
+            serde_json::json!({ "error": "no project root" }).to_string(),
+            "current uploads",
+        );
+    };
+    let session_dir =
+        session_dir.unwrap_or_else(|| crate::web_gateway::pending_upload_session_dir(&root));
+    let result = tokio::task::spawn_blocking(move || {
+        serde_json::to_string(&crate::upload_store::list_uploads(&session_dir, &root))
+            .unwrap_or_else(|_| "[]".to_string())
+    })
+    .await;
+    match result {
+        Ok(body) => json_body_response(id, body, "current uploads"),
+        Err(e) => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": format!("current uploads task failed: {e}"),
+        }),
+    }
+}
+
+pub(crate) async fn api_session_current_upload_task_response(
+    id: String,
+    upload: InboundUploadState,
+    runtime: ControlRuntime,
+) -> ControlTaskResponse {
+    let params = upload.params.clone();
+    let name = optional_string_param(&params, &["name", "filename", "file_name"])
+        .unwrap_or_else(|| "upload.bin".to_string());
+    let mime = optional_string_param(&params, &["mime", "content_type", "contentType"])
+        .unwrap_or_else(|| "application/octet-stream".to_string());
+    let requested_destination = optional_string_param(&params, &["destination"])
+        .as_deref()
+        .and_then(crate::upload_store::UploadDestination::from_str)
+        .unwrap_or(crate::upload_store::UploadDestination::Task);
+    let (session_log, daemon_session_id) = {
+        let session = runtime.shared_session.read().await;
+        (
+            session.session_log.clone(),
+            Some(runtime.session_id.clone()),
+        )
+    };
+    let project_root = runtime.project_root.clone();
+    let bus = runtime.bus.clone();
+    let result = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::current_upload_commit_response_body(
+            project_root.as_deref(),
+            session_log.as_ref(),
+            daemon_session_id.as_deref(),
+            &name,
+            &mime,
+            requested_destination,
+            upload.tmp,
+            upload.received_bytes,
+            &bus,
+        )
+    })
+    .await;
+    let frame = match result {
+        Ok((status, body)) => {
+            http_body_response(id.clone(), status_line_code(status), body, "current upload")
+        }
+        Err(e) => serde_json::json!({
+            "t": "response",
+            "id": id.clone(),
+            "ok": false,
+            "error": format!("upload commit task failed: {e}"),
+        }),
+    };
+    ControlTaskResponse {
+        id,
+        frame,
+        byte_stream: None,
+        done: true,
+    }
+}
+
+pub(crate) async fn api_session_delete_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let session_id = string_param(&params, &["session_id", "sessionId", "id"]);
+    let target =
+        optional_string_param(&params, &["target"]).unwrap_or_else(|| "session".to_string());
+    let result = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::delete_session_data(&session_id, &target)
+    })
+    .await;
+    match result {
+        Ok(body) => json_body_response(id, body, "session delete"),
+        Err(e) => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": format!("session delete task failed: {e}"),
+        }),
+    }
+}
+
+pub(crate) async fn api_session_current_agent_output_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    match active_session_log_dir(runtime).await {
+        Ok(Some(log_dir)) => http_wire_response(
+            id,
+            crate::web_gateway::current_agent_output_post_response(&body_text, &log_dir),
+            "agent output",
+        ),
+        Ok(None) => http_body_response(
+            id,
+            404,
+            serde_json::json!({"error": "no active session log"}).to_string(),
+            "agent output",
+        ),
+        Err(error) => http_body_response(
+            id,
+            500,
+            serde_json::json!({"error": error}).to_string(),
+            "agent output",
+        ),
+    }
+}
+
+pub(crate) async fn api_session_agent_output_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let session_id = string_param(&params, &["session_id", "sessionId", "id"]);
+    if session_id.is_empty() {
+        return serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": "missing session_id",
+        });
+    }
+    let source = string_param(&params, &["source"]).trim().to_string();
+    let source = if source.is_empty() {
+        "intendant".to_string()
+    } else {
+        source
+    };
+    let body_text = params_body_text(Some(&params));
+    let response = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::session_agent_output_post_response(&body_text, &session_id, &source)
+    })
+    .await;
+    match response {
+        Ok(response) => http_wire_response(id, response, "session agent output"),
+        Err(e) => http_body_response(
+            id,
+            500,
+            serde_json::json!({"error": format!("session output task failed: {e}")}).to_string(),
+            "session agent output",
+        ),
+    }
+}
+
+pub(crate) async fn api_session_current_history_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let (file_watcher, _) = active_history_handles(runtime).await;
+    let (status_line, body) = crate::web_gateway::handle_history_get(file_watcher.as_ref()).await;
+    http_body_response(id, status_line_code(status_line), body, "session history")
+}
+
+pub(crate) async fn api_session_current_rollback_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let (file_watcher, agent_state) = active_history_handles(runtime).await;
+    let (status_line, body) = crate::web_gateway::handle_history_rollback(
+        &body_text,
+        file_watcher.as_ref(),
+        agent_state.as_ref(),
+        &runtime.bus,
+    )
+    .await;
+    http_body_response(id, status_line_code(status_line), body, "session rollback")
+}
+
+pub(crate) async fn api_session_current_redo_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let (file_watcher, agent_state) = active_history_handles(runtime).await;
+    let (status_line, body) =
+        crate::web_gateway::handle_history_redo(file_watcher.as_ref(), agent_state.as_ref()).await;
+    http_body_response(id, status_line_code(status_line), body, "session redo")
+}
+
+pub(crate) async fn api_session_current_prune_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let (file_watcher, _) = active_history_handles(runtime).await;
+    let (status_line, body) = crate::web_gateway::handle_history_prune(file_watcher.as_ref()).await;
+    http_body_response(id, status_line_code(status_line), body, "session prune")
+}
+
+pub(crate) async fn api_session_current_changes_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let request_line = changes_request_line(params);
+    let (snapshot_dir, project_root) = active_changes_handles(runtime).await;
+    let home = crate::platform::home_dir();
+    let result = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::handle_changes_request_for_home(
+            &request_line,
+            snapshot_dir.as_deref(),
+            project_root.as_deref(),
+            &home,
+        )
+    })
+    .await;
+    match result {
+        Ok((status_line, body)) => {
+            http_body_response(id, status_line_code(status_line), body, "session changes")
+        }
+        Err(e) => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": format!("session changes task failed: {e}"),
+        }),
+    }
+}
+
+pub(crate) async fn api_session_context_snapshot_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let session_id = string_param(&params, &["session_id", "sessionId", "id"]);
+    if session_id.is_empty() {
+        return missing_param_response(id, "session_id");
+    }
+    let source = optional_string_param(&params, &["source"]).unwrap_or_else(|| "intendant".into());
+    let file = optional_string_param(&params, &["file"]);
+    let request_id = optional_string_param(&params, &["request_id", "requestId"]);
+    let request_index = match optional_u64_param(&params, &["request_index", "requestIndex"]) {
+        Ok(value) => value,
+        Err(error) => {
+            return http_body_response(
+                id,
+                400,
+                serde_json::json!({ "error": error }).to_string(),
+                "context snapshot",
+            );
+        }
+    };
+    let ts = optional_string_param(&params, &["ts"]);
+    let home = crate::platform::home_dir();
+    let result = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::session_context_snapshot_response_body(
+            &home,
+            &session_id,
+            &source,
+            file,
+            request_id,
+            request_index,
+            ts,
+        )
+    })
+    .await;
+    match result {
+        Ok((status_line, body)) => {
+            http_body_response(id, status_line_code(status_line), body, "context snapshot")
+        }
+        Err(e) => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": format!("context snapshot task failed: {e}"),
+        }),
+    }
+}
+
+pub(crate) async fn api_session_current_upload_delete_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let upload_id = string_param(&params, &["upload_id", "uploadId", "id"]);
+    let (project_root, session_dir) = match active_upload_handles(runtime).await {
+        Ok(handles) => handles,
+        Err(error) => {
+            return http_body_response(
+                id,
+                500,
+                serde_json::json!({ "error": error }).to_string(),
+                "upload delete",
+            );
+        }
+    };
+    let result = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::current_upload_delete_response_body(
+            project_root.as_deref(),
+            session_dir.as_deref(),
+            &upload_id,
+        )
+    })
+    .await;
+    match result {
+        Ok((status_line, body, deleted_id)) => {
+            if let Some(id) = deleted_id {
+                runtime
+                    .bus
+                    .send(crate::event::AppEvent::UploadDeleted { id });
+            }
+            http_body_response(id, status_line_code(status_line), body, "upload delete")
+        }
+        Err(e) => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": format!("upload delete task failed: {e}"),
+        }),
+    }
+}
+
+pub(crate) async fn api_sessions_search_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    cancel: CancellationToken,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let query = string_param(&params, &["q", "query"]);
+    let source_filter = string_param(&params, &["source", "source_filter", "sourceFilter"]);
+    let source_filter = if source_filter.is_empty() {
+        "all".to_string()
+    } else {
+        source_filter
+    };
+    let mode = string_param(&params, &["mode"]);
+    let project_filter = control_project_filter(&params);
+    let body = crate::web_gateway::sessions_search_response_body_with_cancel(
+        query,
+        source_filter,
+        mode,
+        project_filter,
+        cancel,
+    )
+    .await;
+    json_body_response(id, body, "session search")
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::*;
+    use crate::dashboard_control::tests::{runtime};
+
+    #[tokio::test]
+    async fn session_report_rpc_returns_zip_for_active_log() {
+        let dir = tempfile::tempdir().unwrap();
+        let session_dir = dir.path().join("session-report");
+        let log = crate::session_log::SessionLog::open(session_dir.clone()).unwrap();
+        std::fs::write(session_dir.join("summary.json"), "{\"ok\":true}\n").unwrap();
+        std::fs::create_dir_all(session_dir.join("turns")).unwrap();
+        std::fs::write(
+            session_dir.join("turns").join("turn_001_stdout.txt"),
+            "hello\n",
+        )
+        .unwrap();
+
+        let rt = runtime();
+        {
+            let mut session = rt.shared_session.write().await;
+            session.session_log = Some(Arc::new(std::sync::Mutex::new(log)));
+        }
+        let report = api_session_report_task_response(
+            "report1".to_string(),
+            Some(&serde_json::json!({})),
+            &rt,
+        )
+        .await;
+        assert!(report.done);
+        assert_eq!(report.id, "report1");
+        assert!(report.byte_stream.is_some());
+        let stream = report.byte_stream.unwrap();
+        assert_eq!(stream.id, "report1");
+        assert_eq!(stream.stream_id, "report1:session-report");
+        assert_eq!(stream.content_type, "application/zip");
+        assert!(stream.filename.as_deref().unwrap_or("").ends_with(".zip"));
+        assert_eq!(stream.result["ok"], true);
+        assert_eq!(stream.result["content_type"], "application/zip");
+        assert!(stream.result["filename"]
+            .as_str()
+            .unwrap_or("")
+            .ends_with(".zip"));
+        assert_eq!(
+            stream.result["size"].as_u64().unwrap(),
+            stream.bytes.len() as u64
+        );
+        let mut zip = zip::ZipArchive::new(std::io::Cursor::new(stream.bytes)).unwrap();
+        assert!(zip.by_name("summary.json").is_ok());
+        assert!(zip.by_name("turns/turn_001_stdout.txt").is_ok());
+
+        let invalid = api_session_report_task_response(
+            "report2".to_string(),
+            Some(&serde_json::json!({ "session_id": "../bad" })),
+            &rt,
+        )
+        .await;
+        assert!(invalid.byte_stream.is_none());
+        assert_eq!(invalid.frame["result"]["_httpStatus"], 400);
+        assert_eq!(invalid.frame["result"]["_httpOk"], false);
+    }
+
+    #[tokio::test]
+    async fn session_delete_rpc_preserves_body_shape() {
+        let invalid_session = api_session_delete_response(
+            "del1".to_string(),
+            Some(&serde_json::json!({ "session_id": "../bad" })),
+        )
+        .await;
+        assert_eq!(invalid_session["t"], "response");
+        assert_eq!(invalid_session["ok"], true);
+        assert_eq!(invalid_session["result"]["ok"], false);
+        assert_eq!(invalid_session["result"]["error"], "invalid session id");
+    }
+
+    #[tokio::test]
+    async fn context_snapshot_rpc_preserves_http_status() {
+        let invalid_session = api_session_context_snapshot_response(
+            "ctx1".to_string(),
+            Some(&serde_json::json!({
+                "session_id": "../bad",
+                "file": "snapshot.json",
+            })),
+        )
+        .await;
+        assert_eq!(invalid_session["t"], "response");
+        assert_eq!(invalid_session["ok"], true);
+        assert_eq!(invalid_session["result"]["error"], "invalid session id");
+        assert_eq!(invalid_session["result"]["_httpStatus"], 400);
+        assert_eq!(invalid_session["result"]["_httpOk"], false);
+
+        let missing_selector = api_session_context_snapshot_response(
+            "ctx2".to_string(),
+            Some(&serde_json::json!({
+                "session_id": "missing-session",
+            })),
+        )
+        .await;
+        assert_eq!(
+            missing_selector["result"]["error"],
+            "missing snapshot selector"
+        );
+        assert_eq!(missing_selector["result"]["_httpStatus"], 400);
+        assert_eq!(missing_selector["result"]["_httpOk"], false);
+
+        let invalid_index = api_session_context_snapshot_response(
+            "ctx3".to_string(),
+            Some(&serde_json::json!({
+                "session_id": "missing-session",
+                "request_index": "abc",
+            })),
+        )
+        .await;
+        assert_eq!(invalid_index["result"]["error"], "invalid request_index");
+        assert_eq!(invalid_index["result"]["_httpStatus"], 400);
+        assert_eq!(invalid_index["result"]["_httpOk"], false);
+    }
+
+    #[tokio::test]
+    async fn current_upload_delete_preserves_http_status() {
+        let rt_no_root = runtime();
+        let no_root = api_session_current_upload_delete_response(
+            "upl1".to_string(),
+            Some(&serde_json::json!({ "id": "missing-upload" })),
+            &rt_no_root,
+        )
+        .await;
+        assert_eq!(no_root["t"], "response");
+        assert_eq!(no_root["ok"], true);
+        assert_eq!(no_root["result"]["error"], "no project root");
+        assert_eq!(no_root["result"]["_httpStatus"], 404);
+        assert_eq!(no_root["result"]["_httpOk"], false);
+
+        let dir = tempfile::tempdir().unwrap();
+        let rt = runtime();
+        {
+            let mut session = rt.shared_session.write().await;
+            session.project_root_for_changes = Some(dir.path().to_path_buf());
+        }
+        let missing_id = api_session_current_upload_delete_response(
+            "upl2".to_string(),
+            Some(&serde_json::json!({})),
+            &rt,
+        )
+        .await;
+        assert_eq!(missing_id["result"]["error"], "missing upload id");
+        assert_eq!(missing_id["result"]["_httpStatus"], 400);
+        assert_eq!(missing_id["result"]["_httpOk"], false);
+    }
+
+    #[tokio::test]
+    async fn current_uploads_lists_pending_uploads() {
+        let project = tempfile::tempdir().unwrap();
+        let mut rt = runtime();
+        rt.project_root = Some(project.path().to_path_buf());
+        {
+            let mut session = rt.shared_session.write().await;
+            session.project_root_for_changes = Some(project.path().to_path_buf());
+        }
+        let bytes = b"dashboard list upload bytes";
+        let tmp = tempfile::NamedTempFile::new().unwrap();
+        std::fs::write(tmp.path(), bytes).unwrap();
+
+        let (status, body) = crate::web_gateway::current_upload_commit_response_body(
+            Some(project.path()),
+            None,
+            Some(rt.session_id.as_str()),
+            "listed.txt",
+            "text/plain",
+            crate::upload_store::UploadDestination::Task,
+            tmp,
+            bytes.len(),
+            &rt.bus,
+        );
+        assert_eq!(status, "200 OK");
+        let descriptor: crate::upload_store::UploadDescriptor =
+            serde_json::from_str(&body).unwrap();
+
+        let response = api_session_current_uploads_response("uploads1".to_string(), &rt).await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["ok"], true);
+        let uploads = response["result"].as_array().expect("uploads array");
+        assert!(
+            uploads.iter().any(|upload| upload["id"] == descriptor.id),
+            "upload list did not include committed descriptor: {response}"
+        );
+    }
+
+    #[tokio::test]
+    async fn current_upload_raw_streams_requested_range() {
+        let project = tempfile::tempdir().unwrap();
+        let mut rt = runtime();
+        rt.project_root = Some(project.path().to_path_buf());
+        let bytes = b"dashboard raw upload bytes";
+        let tmp = tempfile::NamedTempFile::new().unwrap();
+        std::fs::write(tmp.path(), bytes).unwrap();
+
+        let (status, body) = crate::web_gateway::current_upload_commit_response_body(
+            Some(project.path()),
+            None,
+            Some(rt.session_id.as_str()),
+            "raw.txt",
+            "text/plain",
+            crate::upload_store::UploadDestination::Task,
+            tmp,
+            bytes.len(),
+            &rt.bus,
+        );
+        assert_eq!(status, "200 OK");
+        let descriptor: crate::upload_store::UploadDescriptor =
+            serde_json::from_str(&body).unwrap();
+
+        let response = api_session_current_upload_raw_task_response(
+            "raw1".to_string(),
+            Some(&serde_json::json!({
+                "id": descriptor.id,
+                "offset": 10,
+                "length": 6,
+            })),
+            &rt,
+        )
+        .await;
+        assert!(response.done);
+        assert_eq!(response.id, "raw1");
+        assert!(response.byte_stream.is_some());
+        let stream = response.byte_stream.unwrap();
+        assert_eq!(stream.id, "raw1");
+        assert_eq!(stream.stream_id, format!("raw1:upload:{}", descriptor.id));
+        assert_eq!(stream.content_type, "text/plain");
+        assert_eq!(stream.filename.as_deref(), Some("raw.txt"));
+        assert_eq!(stream.bytes, &bytes[10..16]);
+        assert_eq!(stream.result["ok"], true);
+        assert_eq!(stream.result["id"], descriptor.id);
+        assert_eq!(stream.result["name"], "raw.txt");
+        assert_eq!(stream.result["filename"], "raw.txt");
+        assert_eq!(stream.result["mime"], "text/plain");
+        assert_eq!(stream.result["content_type"], "text/plain");
+        assert_eq!(stream.result["size"], 6);
+        assert_eq!(stream.result["total_size"], bytes.len());
+        assert_eq!(stream.result["offset"], 10);
+        assert_eq!(stream.result["range_start"], 10);
+        assert_eq!(stream.result["range_end"], 16);
+        assert_eq!(stream.result["resumable"], true);
+
+        let invalid = api_session_current_upload_raw_task_response(
+            "raw2".to_string(),
+            Some(&serde_json::json!({
+                "id": descriptor.id,
+                "offset": bytes.len() + 1,
+                "length": 1,
+            })),
+            &rt,
+        )
+        .await;
+        assert!(invalid.done);
+        assert!(invalid.byte_stream.is_none());
+        assert_eq!(invalid.frame["t"], "response");
+        assert_eq!(invalid.frame["ok"], true);
+        assert_eq!(invalid.frame["result"]["_httpStatus"], 416);
+        assert_eq!(invalid.frame["result"]["_httpOk"], false);
+        assert_eq!(
+            invalid.frame["result"]["error"],
+            "range start beyond upload size"
+        );
+    }
+
+    #[tokio::test]
+    async fn control_stream_json_lines_emit_lifecycle_frames() {
+        let (line_tx, line_rx) = mpsc::channel::<String>(8);
+        let stream_task = tokio::spawn(async move {
+            for line in [
+                r#"{"type":"start","limit":1,"quick_limit":1}"#,
+                r#"{"type":"session","partial":true,"session":{"session_id":"s1"}}"#,
+                r#"{"type":"done"}"#,
+            ] {
+                line_tx.send(format!("{line}\n")).await.unwrap();
+            }
+        });
+        let (task_tx, mut rx) = mpsc::channel::<ControlTaskResponse>(8);
+
+        stream_json_lines_response(
+            "stream1".to_string(),
+            "api_sessions_stream".to_string(),
+            line_rx,
+            stream_task,
+            task_tx,
+            CancellationToken::new(),
+        )
+        .await;
+
+        let mut frames = Vec::new();
+        while let Some(task) = rx.recv().await {
+            frames.push(task);
+            if frames.last().unwrap().done {
+                break;
+            }
+        }
+
+        assert_eq!(frames.len(), 5);
+        assert_eq!(frames[0].frame["t"], "stream_start");
+        assert_eq!(frames[0].frame["method"], "api_sessions_stream");
+        assert!(!frames[0].done);
+        assert_eq!(frames[1].frame["t"], "stream_event");
+        assert_eq!(frames[1].frame["seq"], 0);
+        assert_eq!(frames[1].frame["event"]["type"], "start");
+        assert_eq!(frames[2].frame["event"]["session"]["session_id"], "s1");
+        assert_eq!(frames[3].frame["event"]["type"], "done");
+        assert_eq!(frames[4].frame["t"], "stream_end");
+        assert_eq!(frames[4].frame["ok"], true);
+        assert_eq!(frames[4].frame["result"]["events"], 3);
+        assert!(frames[4].done);
+    }
+
+    #[test]
+    fn session_rpc_params_parse_limits_and_ids() {
+        assert_eq!(
+            control_session_limit(&serde_json::json!({})),
+            Some(CONTROL_DEFAULT_SESSION_LIMIT)
+        );
+        assert_eq!(
+            control_session_limit(&serde_json::json!({"limit": 25})),
+            Some(25)
+        );
+        assert_eq!(
+            control_session_limit(&serde_json::json!({"limit": 0})),
+            Some(CONTROL_DEFAULT_SESSION_LIMIT)
+        );
+        assert_eq!(
+            control_session_limit(&serde_json::json!({"limit": "nope"})),
+            Some(CONTROL_DEFAULT_SESSION_LIMIT)
+        );
+        assert_eq!(
+            control_session_limit(&serde_json::json!({"limit": "all"})),
+            None
+        );
+        assert_eq!(control_session_detail_limit(&serde_json::json!({})), None);
+        assert_eq!(
+            control_session_detail_limit(&serde_json::json!({"limit": 25})),
+            Some(25)
+        );
+        assert_eq!(
+            control_session_detail_limit(&serde_json::json!({"limit": "25"})),
+            Some(25)
+        );
+        assert_eq!(
+            control_session_detail_limit(&serde_json::json!({"limit": "all"})),
+            None
+        );
+        assert_eq!(
+            control_session_ids(&serde_json::json!({"ids": "a,b, c"})),
+            vec!["a".to_string(), "b".to_string(), "c".to_string()]
+        );
+        assert_eq!(
+            control_session_ids(&serde_json::json!({"ids": ["a,b", "c"]})),
+            vec!["a".to_string(), "b".to_string(), "c".to_string()]
+        );
+        assert_eq!(
+            sessions_stream_request_line(Some(&serde_json::json!({}))),
+            "GET /api/sessions/stream HTTP/1.1"
+        );
+        assert_eq!(
+            sessions_stream_request_line(Some(&serde_json::json!({"limit": "all"}))),
+            "GET /api/sessions/stream?limit=all HTTP/1.1"
+        );
+        assert_eq!(
+            sessions_stream_request_line(Some(&serde_json::json!({"limit": 25}))),
+            "GET /api/sessions/stream?limit=25 HTTP/1.1"
+        );
+        assert_eq!(
+            control_project_filter(&serde_json::json!({"projects": ["a", " b "]})),
+            vec!["a".to_string(), "b".to_string()]
+        );
+        assert_eq!(
+            control_project_filter(&serde_json::json!({"projects": "[\"a\",\"b\"]"})),
+            vec!["a".to_string(), "b".to_string()]
+        );
+        assert_eq!(
+            control_project_filter(&serde_json::json!({"projects": "a,b"})),
+            vec!["a".to_string(), "b".to_string()]
+        );
+        assert_eq!(
+            control_capability_query(
+                &serde_json::json!({"capabilities": ["display", "custom:gpu"]})
+            ),
+            "capability=display&capability=custom:gpu"
+        );
+    }
+}
